[R] refactoring by revue
